+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 0.7.6dev
 
+* [Fix] Fix error when checking if custom connection was PEP 249 Compliant (#517)
 * [Feature] Support for printing capture variables using `=<<` syntax (by [@jorisroovers](https://github.com/jorisroovers))
 * [Feature] Adds `--persist-replace` argument to replace existing tables when persisting data frames (#440)
+
 
 ## 0.7.5 (2023-05-24)